refactor: rename stored payload module
