+++ storage/src/record/mod.rs
@@ -2,7 +2,7 @@ mod batcher;
 mod encoding;
 mod encryption;
 mod iterator;
-mod stream_payload;
+mod stored_payload;
 
 pub use batcher::{RecordBatch, RecordBatcher};
 use bytes::{Buf, BufMut, Bytes, BytesMut};
@@ -18,7 +18,7 @@ use s2_common::{
     encryption::EncryptionAlgorithm,
     record::{CommandRecord, Metered, MeteredSize, Record, SeqNum, Sequenced},
 };
-pub use stream_payload::{
+pub use stored_payload::{
     StoredAppendInput, StoredAppendRecord, StoredAppendRecordBatch, StoredAppendRecordParts,
     StoredReadBatch, StoredReadSessionOutput, decrypt_read_session_output, encrypt_append_input,
 };
