+++ storage/src/record/stored_payload.rs
